refactor(desktop): extract agent studio diff polling controller
